:zap: create project fetcher as hook and use it in dashboard cards component
