+++ frontend/src/components/DashboardCards/DashboardCards.jsx
@@ -1,23 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import DCard from "./DCrad/DCard";
 import "./DashboardCards.css";
 import CustomSelection from "../CustomSelection/CustomSelection";
-import request , {FETCH} from '../../utils/requests';
-import { usePermission , MANAGER} from "../../Hooks";
+import {useProjectsFetcher} from "../../Hooks/fetchers";
 
 const DashboardCards = (props) => {
-    const {permission} = usePermission()
-    const [projects , setProjects] = useState([]);
-    const fetchProjects = async()=>{
-        const {data} =  await request("/projects" , FETCH , [false,false])
-        setProjects(data)
-    }
-    useEffect(()=>{
-        if (permission.role === MANAGER){
-            fetchProjects();
-        }
-    } , [])
-
+    const {projects} = useProjectsFetcher()
     return (
         <div className="dashboard-cards">
             <div className="row">
